refactor: separate formula text editing from release I/O
