+++ tests/test_update_formula.py
@@ -86,18 +86,18 @@ class UpdateFormulaTest(unittest.TestCase):
                     update_formula,
                     sha256=mock.DEFAULT,
                     verify_remote_source_tag=mock.DEFAULT,
-                    seed_formula=mock.DEFAULT,
                     update_cask=mock.DEFAULT,
                 ) as operations,
+                mock.patch.object(update_formula.formula_text, "seed_formula") as seed,
                 mock.patch.object(update_formula.urllib.request.OpenerDirector, "open") as network,
                 mock.patch.object(update_formula.subprocess, "run") as process,
                 mock.patch.object(pathlib.Path, "write_text") as write,
             ):
-                for operation in (*operations.values(), network, process, write):
+                for operation in (*operations.values(), seed, network, process, write):
                     operation.side_effect = AssertionError("side effect before obsolete-contract rejection")
                 with self.assertRaisesRegex(SystemExit, "Crabbox.*ordinary.*assets.*docs/RELEASING.md"):
                     update_formula.main(arguments)
-                for operation in (*operations.values(), network, process, write):
+                for operation in (*operations.values(), seed, network, process, write):
                     operation.assert_not_called()
         finally:
             os.chdir(previous_directory)
@@ -145,8 +145,8 @@ class UpdateFormulaTest(unittest.TestCase):
         original = (ROOT / "Formula" / "crabbox.rb").read_text()
         assets = crabbox_assets()
         expected = original
-        for match in update_formula.iter_url_sha_pairs(original):
-            target = update_formula.classify_target(match.group("url"), {}, "1.2.3")
+        for match in update_formula.formula_text.iter_url_sha_pairs(original):
+            target = update_formula.formula_text.classify_target(match.group("url"), {}, "1.2.3")
             item = assets[target]
             expected = expected.replace(match.group("url"), update_formula.explicit_asset_url(
                 "openclaw/crabbox", "v1.2.3", item["name"],
@@ -254,7 +254,7 @@ class UpdateFormulaTest(unittest.TestCase):
                 self.assertEqual(network.call_count, 4)
             finally:
                 os.chdir(previous_directory)
-            pairs = update_formula.iter_url_sha_pairs(path.read_text())
+            pairs = update_formula.formula_text.iter_url_sha_pairs(path.read_text())
             self.assertCountEqual(
                 [(match.group("url"), match.group("sha")) for match in pairs],
                 [(update_formula.explicit_asset_url("openclaw/crabbox", "v1.2.3", item["name"]), item["sha256"])
@@ -275,7 +275,7 @@ class UpdateFormulaTest(unittest.TestCase):
                     with (
                         mock.patch.object(update_formula, "verify_remote_source_tag") as verify_tag,
                         mock.patch.object(update_formula, "sha256") as download,
-                        mock.patch.object(update_formula, "seed_formula") as seed,
+                        mock.patch.object(update_formula.formula_text, "seed_formula") as seed,
                         mock.patch.object(update_formula, "update_cask") as cask,
                         mock.patch.object(pathlib.Path, "write_text") as write,
                     ):
@@ -378,7 +378,7 @@ class UpdateFormulaTest(unittest.TestCase):
         major, minor, patch = version_match.groups()
         version = f"{major}.{minor}.{int(patch) + 1}"
         hashes = {target: str(index) * 64 for index, target in enumerate(update_formula.RELEASE_TARGETS, 1)}
-        old_pairs = {(match.group("url"), match.group("sha")) for match in update_formula.iter_url_sha_pairs(formula)}
+        old_pairs = {(match.group("url"), match.group("sha")) for match in update_formula.formula_text.iter_url_sha_pairs(formula)}
         metadata = r'(?m)^\s*(?:version|url|sha256) "[^"\n]+"$'
 
         for mode in ("explicit-assets", "legacy-template", "verified-hashes"):
@@ -433,7 +433,7 @@ class UpdateFormulaTest(unittest.TestCase):
                     self.assertCountEqual([call.args[0] for call in download.call_args_list], expected)
                 pairs = [
                     (match.group("url").replace("#{version}", version), match.group("sha"))
-                    for match in update_formula.iter_url_sha_pairs(updated)
+                    for match in update_formula.formula_text.iter_url_sha_pairs(updated)
                 ]
                 self.assertCountEqual(pairs, expected.items())
                 self.assertTrue(set(pairs).isdisjoint(old_pairs))
@@ -878,7 +878,7 @@ end
             "linux_arm64": "4" * 64,
         }
 
-        updated = update_formula.render_verified_target_formula(
+        updated = update_formula.formula_text.render_verified_target_formula(
             formula,
             "openclaw/example",
             "v0.36.1",
@@ -920,7 +920,7 @@ end
         )
         for description, candidate, message in cases:
             with self.subTest(description=description), self.assertRaisesRegex(SystemExit, message):
-                update_formula.render_verified_target_formula(
+                update_formula.formula_text.render_verified_target_formula(
                     candidate,
                     "openclaw/example",
                     "v0.43.1",
@@ -940,7 +940,7 @@ end
             "linux_arm64": "4" * 64,
         }
 
-        updated = update_formula.render_verified_target_formula(
+        updated = update_formula.formula_text.render_verified_target_formula(
             formula,
             "openclaw/wacli",
             "v0.12.1",
@@ -965,7 +965,7 @@ end
             self.assertEqual(updated.count(digest), 1)
 
     def test_seed_formula_escapes_ruby_description(self) -> None:
-        seeded = update_formula.seed_formula(
+        seeded = update_formula.formula_text.seed_formula(
             "example",
             "openclaw/example",
             "1.2.3",
@@ -1059,7 +1059,7 @@ end
 end
 '''
 
-        updated = update_formula.update_url_and_sha_in_stanza(
+        updated = update_formula.formula_text.update_url_and_sha_in_stanza(
             text,
             "on_linux",
             "https://github.com/steipete/camsnap/archive/refs/tags/v0.3.0.tar.gz",
@@ -1071,7 +1071,7 @@ end
         self.assertEqual(updated.count("cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc"), 2)
 
     def test_target_update_handles_version_length_change(self) -> None:
-        formula = update_formula.seed_formula(
+        formula = update_formula.formula_text.seed_formula(
             "example",
             "openclaw/example",
             "0.7.9",
@@ -1108,14 +1108,14 @@ end
             "https://github.com/openclaw/example/releases/download/v0.43.0/"
             "example_0.43.0_mystery.tar.gz"
         )
-        self.assertIsNone(update_formula.classify_target(mystery_url, {}, "0.43.0"))
+        self.assertIsNone(update_formula.formula_text.classify_target(mystery_url, {}, "0.43.0"))
         formula = platform_install_formula().replace(
             '  license "MIT"',
             '  version "0.43.0"\n  license "MIT"',
         ).replace("example_0.43.0_linux_arm64.tar.gz", "example_0.43.0_mystery.tar.gz")
         classified = [
-            update_formula.classify_target(match.group("url"), {}, "0.43.0")
-            for match in update_formula.iter_url_sha_pairs(formula)
+            update_formula.formula_text.classify_target(match.group("url"), {}, "0.43.0")
+            for match in update_formula.formula_text.iter_url_sha_pairs(formula)
         ]
         self.assertEqual(classified.count(None), 1)
         self.assertEqual(len([target for target in classified if target]), 3)
@@ -1304,7 +1304,7 @@ end
 '''
 
         with self.assertRaises(SystemExit) as raised:
-            update_formula.update_url_and_sha_in_stanza(
+            update_formula.formula_text.update_url_and_sha_in_stanza(
                 text,
                 "on_linux",
                 "https://github.com/steipete/example/archive/refs/tags/v1.0.1.tar.gz",
@@ -1345,7 +1345,7 @@ end
 end
 '''
 
-        self.assertTrue(update_formula.uses_stanza_url_mode(text, "0.9.2"))
+        self.assertTrue(update_formula.formula_text.uses_stanza_url_mode(text, "0.9.2"))
 
     def test_converts_duplicate_platform_stanzas_to_target_urls(self) -> None:
         text = '''class Wacli < Formula
@@ -1380,7 +1380,7 @@ end
 end
 '''
 
-        updated = update_formula.convert_stanza_url_mode_to_targets(
+        updated = update_formula.formula_text.convert_stanza_url_mode_to_targets(
             text,
             "openclaw/wacli",
             "v0.9.3",
@@ -1417,7 +1417,7 @@ end
 end
 '''
 
-        updated = update_formula.insert_target_stanzas(
+        updated = update_formula.formula_text.insert_target_stanzas(
             text,
             "steipete/sag",
             "v0.3.1",
@@ -1444,8 +1444,8 @@ end
 end
 '''
 
-        updated = update_formula.update_version(text, "0.27.0")
-        updated = update_formula.update_top_level_url_and_sha(
+        updated = update_formula.formula_text.update_version(text, "0.27.0")
+        updated = update_formula.formula_text.update_top_level_url_and_sha(
             updated,
             "https://github.com/steipete/CodexBar/releases/download/v0.27.0/CodexBar-macos-universal-0.27.0.zip",
             "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
